@@ -4,9 +4,9 @@
     "description": "Sylius Adyen payment gateway plugin.",
     "license": "MIT",
     "require": {
-        "php": "^7.3",
-        "sylius/sylius": "^1.7",
-        "adyen/php-api-library": "^8.0"
+        "php": "^7.4 || ^8.0",
+        "sylius/sylius": "^1.10",
+        "adyen/php-api-library": "^10.1"
     },
     "require-dev": {
         "behat/behat": "^3.6.1",
@@ -15,25 +15,28 @@
         "dmore/chrome-mink-driver": "^2.7",
         "friends-of-behat/mink": "^1.8",
         "friends-of-behat/mink-browserkit-driver": "^1.4",
+        "friends-of-behat/mink-debug-extension": "^2.0.0",
         "friends-of-behat/mink-extension": "^2.4",
         "friends-of-behat/page-object-extension": "^0.3",
         "friends-of-behat/suite-settings-extension": "^1.0",
         "friends-of-behat/symfony-extension": "^2.1",
         "friends-of-behat/variadic-extension": "^1.3",
-        "lakion/mink-debug-extension": "^1.2.3",
-        "phpspec/phpspec": "^6.1",
-        "phpstan/phpstan": "0.12.25",
-        "phpstan/phpstan-doctrine": "0.12.13",
-        "phpstan/phpstan-webmozart-assert": "0.12.4",
-        "phpunit/phpunit": "^8.5",
+        "friendsofsymfony/oauth-server-bundle": "^1.6 || >2.0.0-alpha.0 ^2.0@dev",
+        "phpspec/phpspec": "^7.0",
+        "phpstan/extension-installer": "^1.0",
+        "phpstan/phpstan": "0.12.85",
+        "phpstan/phpstan-doctrine": "0.12.33",
+        "phpstan/phpstan-strict-rules": "^0.12.0",
+        "phpstan/phpstan-webmozart-assert": "0.12.12",
+        "phpunit/phpunit": "^9.5",
         "sensiolabs/security-checker": "^6.0",
-        "sylius-labs/coding-standard": "^3.1",
-        "symfony/browser-kit": "^4.4",
-        "symfony/debug-bundle": "^4.4|^5.0",
-        "symfony/dotenv": "^4.4|^5.0",
-        "symfony/intl": "^4.4|^5.0",
-        "symfony/web-profiler-bundle": "^4.4|^5.0",
-        "symfony/web-server-bundle": "^4.4|^5.0"
+        "sylius-labs/coding-standard": "^4.0",
+        "symfony/browser-kit": "^4.4 || ^5.2",
+        "symfony/debug-bundle": "^4.4 || ^5.2",
+        "symfony/dotenv": "^4.4 || ^5.2",
+        "symfony/intl": "^4.4 || ^5.2",
+        "symfony/web-profiler-bundle": "^4.4 || ^5.2",
+        "vimeo/psalm": "4.7.1"
     },
     "autoload": {
         "psr-4": {
